# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AtomQuest Goal Setting & Tracking Portal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users
      parameters:
        - name: role
          in: query
          schema:
            type: string
            enum: [employee, manager, admin]
        - name: managerId
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found

  # Cycles
...
  /goal-sheets:
    get:
      operationId: listGoalSheets
      tags: [goal-sheets]
      summary: List goal sheets (role-filtered)
      parameters:
        - name: cycleId
          in: query
          schema:
            type: ["integer", "null"]
        - name: employeeId
          in: query
          schema:
            type: ["integer", "null"]
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, submitted, approved, returned]
      responses:
        "200":
          description: List of goal sheets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GoalSheet"
    post:
      operationId: createGoalSheet
      tags: [goal-sheets]
      summary: Create a goal sheet
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoalSheetInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GoalSheet"
...
  /goal-sheets/{sheetId}/goals:
    post:
      operationId: createGoal
      tags: [goals]
      summary: Add a goal to a sheet
      parameters:
        - name: sheetId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoalInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"
  /goals/{id}:
    patch:
      operationId: updateGoal
      tags: [goals]
      summary: Update a goal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoalUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"
    delete:
      operationId: deleteGoal
      tags: [goals]
      summary: Delete a goal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Shared Goals
  /shared-goals:
    get:
      operationId: listSharedGoals
      tags: [shared-goals]
      summary: List shared goals
      responses:
        "200":
          description: List of shared goals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SharedGoal"
    post:
      operationId: createSharedGoal
      tags: [shared-goals]
      summary: Push shared goal to multiple employees
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SharedGoalInput"
      responses:
        "201":
          description: Created and distributed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SharedGoal"
  /shared-goals/{id}:
    patch:
      operationId: updateSharedGoal
      tags: [shared-goals]
      summary: Update shared goal weightage
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SharedGoalUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SharedGoal"

  # Check-ins
...
  /reports/achievement:
    get:
      operationId: getAchievementReport
      tags: [reports]
      summary: Achievement report (Planned vs Actual)
      parameters:
        - name: cycleId
          in: query
          schema:
            type: ["integer", "null"]
        - name: quarter
          in: query
          schema:
            type: string
            enum: [Q1, Q2, Q3, Q4]
        - name: format
          in: query
          schema:
            type: string
            enum: [json, csv]
      responses:
        "200":
          description: Achievement report
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AchievementReportRow"

  # Audit Logs
  /audit-logs:
    get:
      operationId: listAuditLogs
      tags: [audit-logs]
      summary: List audit log entries
      parameters:
        - name: entityType
          in: query
          schema:
            type: string
        - name: entityId
          in: query
          schema:
            type: ["integer", "null"]
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: Audit log entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AuditLog"
components:
  schemas:
...
    User:
      type: object
      required: [id, name, email, role, department]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [employee, manager, admin]
        department:
          type: string
        designation:
          type: ["string", "null"]
        managerId:
          type: ["integer", "null"]
        managerName:
          type: ["string", "null"]
        createdAt:
          type: string

    # Goal Cycle
...
    GoalSheet:
      type: object
      required: [id, employeeId, employeeName, cycleId, cycleName, status, totalWeightage, goalsCount]
      properties:
        id:
          type: integer
        employeeId:
          type: integer
        employeeName:
          type: string
        employeeDepartment:
          type: string
        cycleId:
          type: integer
        cycleName:
          type: string
        status:
          type: string
          enum: [draft, submitted, approved, returned]
        totalWeightage:
          type: number
        goalsCount:
          type: integer
        submittedAt:
          type: ["string", "null"]
        approvedAt:
          type: ["string", "null"]
        managerComment:
          type: ["string", "null"]
        isLocked:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
...
    GoalSheetInput:
      type: object
      required: [cycleId]
      properties:
        cycleId:
          type: integer
...
    Goal:
      type: object
      required: [id, goalSheetId, thrustAreaId, thrustAreaName, title, description, uomType, target, weightage, status]
      properties:
        id:
          type: integer
        goalSheetId:
          type: integer
        thrustAreaId:
          type: integer
        thrustAreaName:
          type: string
        title:
          type: string
        description:
          type: string
        uomType:
          type: string
          enum: [numeric_min, numeric_max, timeline, zero]
        uomUnit:
          type: ["string", "null"]
        target:
          type: string
        weightage:
          type: number
        status:
          type: string
          enum: [not_started, on_track, completed]
        isShared:
          type: boolean
        sharedGoalId:
          type: ["integer", "null"]
        isReadOnly:
          type: boolean
        progressScore:
          type: ["number", "null"]
        createdAt:
          type: string
    GoalInput:
      type: object
      required: [thrustAreaId, title, description, uomType, target, weightage]
      properties:
        thrustAreaId:
          type: integer
        title:
          type: string
        description:
          type: string
        uomType:
          type: string
          enum: [numeric_min, numeric_max, timeline, zero]
        uomUnit:
          type: string
        target:
          type: string
        weightage:
          type: number
    GoalUpdate:
      type: object
      properties:
        thrustAreaId:
          type: integer
        title:
          type: string
        description:
          type: string
        uomType:
          type: string
          enum: [numeric_min, numeric_max, timeline, zero]
        uomUnit:
          type: string
        target:
          type: string
        weightage:
          type: number
        status:
          type: string
          enum: [not_started, on_track, completed]

    # Shared Goals
    SharedGoal:
      type: object
      required: [id, title, thrustAreaId, thrustAreaName, target, uomType, cycleId, ownerEmployeeId, recipientCount]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        thrustAreaId:
          type: integer
        thrustAreaName:
          type: string
        target:
          type: string
        uomType:
          type: string
          enum: [numeric_min, numeric_max, timeline, zero]
        cycleId:
          type: integer
        ownerEmployeeId:
          type: integer
        ownerName:
          type: string
        recipientCount:
          type: integer
        createdAt:
          type: string
    SharedGoalInput:
      type: object
      required: [title, thrustAreaId, target, uomType, cycleId, recipientEmployeeIds, defaultWeightage]
      properties:
        title:
          type: string
        description:
          type: string
        thrustAreaId:
          type: integer
        target:
          type: string
        uomType:
          type: string
          enum: [numeric_min, numeric_max, timeline, zero]
        cycleId:
          type: integer
        recipientEmployeeIds:
          type: array
          items:
            type: integer
        defaultWeightage:
          type: number
    SharedGoalUpdate:
      type: object
      properties:
        weightage:
          type: number

    # Check-in
...
    AchievementReportRow:
      type: object
      required: [employeeName, department, goalTitle, thrustArea, uomType, target, weightage, quarter, achievement, progressScore, status]
      properties:
        employeeName:
          type: string
        department:
          type: string
        goalTitle:
          type: string
        thrustArea:
          type: string
        uomType:
          type: string
        target:
          type: string
        weightage:
          type: number
        quarter:
          type: string
        achievement:
          type: ["string", "null"]
        progressScore:
          type: ["number", "null"]
        status:
          type: string

    # Audit Log
    AuditLog:
      type: object
      required: [id, userId, userName, action, entityType, entityId, timestamp]
      properties:
        id:
          type: integer
        userId:
          type: integer
        userName:
          type: string
        action:
          type: string
        entityType:
          type: string
        entityId:
          type: integer
        oldValue:
          type: ["string", "null"]
        newValue:
          type: ["string", "null"]
        reason:
          type: ["string", "null"]
        timestamp:
          type: string